+++ integration_tests/src/tests/mod.rs
@@ -18,7 +18,8 @@ mod extend_sectors_test;
 pub use extend_sectors_test::*;
 mod market_miner_withdrawal_test;
 pub use market_miner_withdrawal_test::*;
-// mod multisig_test;
+mod multisig_test;
+pub use multisig_test::*;
 // mod power_scenario_tests;
 // mod publish_deals_test;
 // mod replica_update_test;
